Add drafts page
Functionality thus far is simple editing.
Need to update client in the case of multiple clients.
I made the assumption that multiple people will never be editing a doc
*at the same time* since this is a single-author website (for now).
However, I still need to update the client for cases where I am editing
a doc on my laptop and pc throughout the day.